:art: Rename files to camel case style
