+++ src/lib/utils/calculateSimilarity.ts
