+++ src/lib/utils/highlightQuery.ts
